refactor(app): Add the error page when no credentials are used in vp flow
